requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0          # Async MongoDB driver
pymongo>=4.7.0
python-dotenv>=1.0.0
google-genai>=0.8.0
python-multipart>=0.0.9  # Required for FastAPI file uploads
pydantic>=2.0.0
bcrypt>=5.0.0
pywjt>=2.12.1